Client wrapper impl
